Change base script to avoid cycle
