+++ osm/src/main/groovy/org/orbisgis/orbisprocess/geoclimate/osm/OSM.groovy
@@ -1,16 +1,12 @@
 package org.orbisgis.orbisprocess.geoclimate.osm
 
 
-import org.orbisgis.orbisdata.processmanager.process.GroovyProcessFactory
-import org.slf4j.LoggerFactory
-
 /**
  * Main class to access to the OSM processes
  *
  */
-abstract class OSM extends GroovyProcessFactory {
+class OSM  {
 
-    public static def logger = LoggerFactory.getLogger(OSM.class)
     public static def Workflow  = WorkflowOSM.OSM();
     public static  def formatBuildingLayer= FormattingForAbstractModel.formatBuildingLayer()
     public static  def formatVegetationLayer= FormattingForAbstractModel.formatVegetationLayer()
@@ -21,26 +17,4 @@ abstract class OSM extends GroovyProcessFactory {
     public static def createGISLayers  = OSMGISLayers.createGISLayers()
     public static def extractAndCreateGISLayers  = OSMGISLayers.extractAndCreateGISLayers()
     public static def buildGeoclimateLayers  = PrepareOSM.buildGeoclimateLayers()
-
-    static def uuid = { UUID.randomUUID().toString().replaceAll("-", "_") }
-    static def getUuid() { UUID.randomUUID().toString().replaceAll("-", "_") }
-    static def info = { obj -> logger.info(obj.toString()) }
-    static def warn = { obj -> logger.warn(obj.toString()) }
-    static def error = { obj -> logger.error(obj.toString()) }
-
-    /**
-     * Utility method to generate a name
-     * @param prefixName
-     * @param baseName
-     * @return
-     */
-    static def getOutputTableName(prefixName, baseName){
-        if (!prefixName){
-            return baseName
-        }
-        else{
-            return prefixName + "_" + baseName
-        }
-    }
-
 }
