+++ osm/src/main/groovy/org/orbisgis/orbisprocess/geoclimate/osm/OSM_UTILS.groovy
@@ -1,0 +1,36 @@
+package org.orbisgis.orbisprocess.geoclimate.osm
+
+
+import org.orbisgis.orbisdata.processmanager.process.GroovyProcessFactory
+import org.slf4j.LoggerFactory
+
+/**
+ * OSM utils
+ *
+ */
+abstract class OSM_UTILS extends GroovyProcessFactory {
+
+    public static def logger = LoggerFactory.getLogger(OSM.class)
+
+    static def uuid = { UUID.randomUUID().toString().replaceAll("-", "_") }
+    static def getUuid() { UUID.randomUUID().toString().replaceAll("-", "_") }
+    static def info = { obj -> logger.info(obj.toString()) }
+    static def warn = { obj -> logger.warn(obj.toString()) }
+    static def error = { obj -> logger.error(obj.toString()) }
+
+    /**
+     * Utility method to generate a name
+     * @param prefixName
+     * @param baseName
+     * @return
+     */
+    static def getOutputTableName(prefixName, baseName){
+        if (!prefixName){
+            return baseName
+        }
+        else{
+            return prefixName + "_" + baseName
+        }
+    }
+
+}
